Working...

+++ clients/java/java-springboot/src/main/java/xcarpaccio/OrderCalculator.java
@@ -6,7 +6,7 @@ public class OrderCalculator {
     public static final double STANDARD_REDUCTION_10K_PLUS = 1 - 10 / 100;
     public static final double STANDARD_REDUCTION_50K_PLUS = 1 - 15 / 100;
 
-    public static final int STANDARD_REDUCTION_THRESHOLD_50K = 50000;
+    public static final double STANDARD_REDUCTION_THRESHOLD_50K = 50000;
 
     public OrderCalculator() {
     }
